requirements.txt
annotated-types
APScheduler
bcrypt
bidict
blinker
cachelib
certifi
charset-normalizer
click
dnspython
Flask
Flask-Bcrypt
Flask-JWT-Extended
Flask-Login
Flask-PyMongo
Flask-Session
Flask-SocketIO
gunicorn
h11
idna
itsdangerous
Jinja2
joblib
MarkupSafe
numpy
packaging
pydantic
pydantic_core
PyJWT
pymongo
python-dotenv
python-engineio
python-socketio
requests
scikit-learn
scipy
simple-websocket
six
threadpoolctl
typing_extensions
tzlocal
urllib3
websocket-client
Werkzeug
wsproto
